trying to execute on linux

+++ main/test/main.cpp
@@ -2,19 +2,19 @@
 #include <fstream>
 #include <algorithm>
 
-void Bubble(int* arr, size_t size)
+void Bubble(int* arr, int size)
 {
-	for (size_t i = size; i > 0; --i)
-		for (size_t j = 1; j < i; ++j)
+	for (int i = size; i > 0; --i)
+		for (int j = 1; j < i; ++j)
 			if (arr[j] < arr[j - 1])
 				std::swap(arr[j], arr[j - 1]);
 }
 
-void Insert(int* arr, size_t size)
+void Insert(int* arr, int size)
 {
-	for (size_t i = 1; i < size; ++i)
+	for (int i = 1; i < size; ++i)
 	{
-		size_t j = i;
+		int j = i;
 		while (j > 0 && arr[j] < arr[j - 1])
 		{
 			std::swap(arr[j], arr[j - 1]);
@@ -23,7 +23,7 @@ void Insert(int* arr, size_t size)
 	}
 }
 
-void MakeHeap(int arr[], size_t length)
+void MakeHeap(int arr[], int length)
 {
 	for (int i = 1; i < length; ++i)
 	{
@@ -42,7 +42,7 @@ void MakeHeap(int arr[], size_t length)
 	}
 }
 
-void SortHeap(int arr[], size_t length)
+void SortHeap(int arr[], int length)
 {
 	for (int i = length - 1; i > 0; --i)
 	{
@@ -88,10 +88,10 @@ int main(int argc, const char* argv[])
 	int type = std::atoi(argv[1]);
 	setlocale(LC_ALL,"Rus");
 	std::ifstream inFile(argv[2]);
-	size_t size = 0;
+	int size = 0;
 	inFile >> size;
 	auto arr = new int[size];
-	for (size_t i = 0; i < size; ++i)
+	for (int i = 0; i < size; ++i)
 	{
 		inFile >> arr[i];
 	}
@@ -126,7 +126,7 @@ int main(int argc, const char* argv[])
 	std::ofstream outFile(argv[3]);
 
 	outFile << size << std::endl;
-	for (size_t i = 0; i < size; ++i)
+	for (int i = 0; i < size; ++i)
 	{
 		outFile << arr[i] << " ";
 	}
